Initial commit with basic functionality, saving data using SQLite
This first version checks a subtitles RSS and cross the entries
with a list of TV shows I watch. If it gets a match, it tries to
find a torrent magnet link, saves it and downloads the subtitle.

The next version should be integrated with Transmission, to
automate the downloads.